Option to turn off all full buildings of a certain type
